ai-dashboard-generator/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pandas==2.2.2
python-multipart==0.0.9
httpx==0.27.0
python-dotenv==1.0.1
langchain==0.2.6
langchain-ollama==0.1.1
chromadb==0.5.4
plotly==5.24.1
